+++ front/src/modules/auth/components/Title.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
 
-import { AnimatedTextWord } from '@/ui/animation/components/AnimatedTextWord';
+import { AnimatedEaseIn } from '../../ui/animation/components/AnimatedEaseIn';
 
 type Props = React.PropsWithChildren & {
   animate?: boolean;
@@ -11,17 +11,17 @@ const StyledTitle = styled.div`
   color: ${({ theme }) => theme.font.color.primary};
   font-size: ${({ theme }) => theme.font.size.xl};
   font-weight: ${({ theme }) => theme.font.weight.semiBold};
-`;
-
-const StyledAnimatedTextWord = styled(AnimatedTextWord)`
-  color: ${({ theme }) => theme.font.color.primary};
-  font-size: ${({ theme }) => theme.font.size.xl};
-  font-weight: ${({ theme }) => theme.font.weight.semiBold};
+  margin-bottom: ${({ theme }) => theme.spacing(4)};
+  margin-top: ${({ theme }) => theme.spacing(4)};
 `;
 
 export function Title({ children, animate = false }: Props) {
-  if (animate && typeof children === 'string') {
-    return <StyledAnimatedTextWord text={children} />;
+  if (animate) {
+    return (
+      <StyledTitle>
+        <AnimatedEaseIn>{children}</AnimatedEaseIn>
+      </StyledTitle>
+    );
   }
 
   return <StyledTitle>{children}</StyledTitle>;
