+++ front/src/modules/auth/sign-in-up/components/FooterNote.tsx
