+++ front/src/modules/auth/sign-in-up/components/HorizontalSeparator.tsx
